Use single instead of double quotes in TNFilterSpec
Not a big deal to me, but it is to my editor vscode: it both squiggle highlights every unnecessecary double quote (recommending single quotes on hover) and auto-switches to single quotes on save. I could work around all that, but vscode is the unofficial official project editor and single quotes is standard throughout the rest of the code base (and ruby). Ripping off the bandage... [I checked that there are no '"...#{...}..."' usages in this file (though it would be a serious bug if vscode didn't handle that).]