Handle the columns count == 0 or 1 cases of SimpleMapper
AFAICT when you only have one column of data, SimpleMappr only reads the latitudes and then gets a 'bad data' error because that doesn't make sense. :(

The second 'empty' column added here does show up as the "legend" title of a second (empty) layer in the SimpleMappr UI, but it never shows on the UI map or saved map in any way.