Fix destroy handling for DWCO-hooked models
The issue with the previous version was that in after_destroy, the destroyed record is no longer in the database, so when you call dwc_occurrences on it, no DWCOs are found.

Those DWCO records need to be found *before* the hooked object is gone, but processed *after* the hooked object is gone.