Change back to after_save_commit for DWCO-hooked model hook
The concern I have is: why is rails running that second save on nested_attributes commits? Presuming it's not frivolous, then someday we may create a model where the result of that second save is required in order to have correct DwCO data, and then the current behavior (run our hook just on the first after_save call that has saved_changes) will be wrong. It's not clear to me if saved_changes is *always* empty on the nested/second save, even if data was saved.

That being the case, we could switch to using `after_create` without `if :saved_changes?` and `after_update, if: :saved_changes?` to restore the saved_changes check when it still(?) works. But I think there's a bigger reason to stick with after_save_commit actually:
* `after_save_commit` runs *after* all other after_save callbacks have completed, which means associated DwCOs will pick up whatever data might have been added in those after_saves. i.e. using after_save_commit eases some of the burden of ensuring that this hook runs after all other dwco-related after-save hooks.