Fix css for multiple leads+items; cleanup #4287
Much simpler than before: this only works because the LeadItems divs for the various leads *all have the same height*, which in turn means all of the leads have the same height since the total height of the columns is always the same.

There were issues with the grid version aside from being more complicated.