Change ui delete of lead_item to delete in all lead descendants #4287
It seems to me if that's not what the user intends then it's a mistake, so we might as well be doing it for them.

The use case is: oops, I meant for this otu to be in a different child of this couplet's parent, so I delete in this branch and add it in that branch.