Adjust where/when lead_items are saved, part 1 #4287
Previously I had tried to keep track of lead_items throughout the entire key, for things like:
1) going to any couplet and being able to see which otus were selected for each lead
2) being able to go back in a key and change your mind about which lead an otu belongs to
3) being able to delete subtrees/couplets and still know which otus were associated with that subtree/couplet

That's actually pretty unwieldy for the variety of operations we support on keys, and was likely never what @proceps had in mind. Per updated discussion we've switched to only keeping otu lists on leaf nodes.

The only one of the above that we really lose doing it this way is the ability to go back in a key and change your mind about an otu/otus (#2). Now you'll have to completely destroy the subtree(s) from where you want to change your mind and start over from there.

Number 1 will no longer appear "for free": if we ever implement it it will probably be behind a button, since it's now more expensive.

Number 3 is now implemented through marginally more expensive queries, not a big deal.