Add remaining lead_item otus count to lead futures list #4287
Unfortunate that this is an extra query for every future lead, even if the key uses no lead_items.

Consider pre-loading a la this discussion: https://github.com/ClosureTree/closure_tree?tab=readme-ov-file#eager-loading, though none of that quite applies directly. Also of course this is kind of only a "big" deal for large keys when you're displaying leads at the top of the key; the further down any key you go the cheaper this is.