Make store.loading a ref-counted spinner for leads #4287
Current thinking: I think a full-screen spinner each time you check or uncheck an otu in the otu list is too much/too slow, so using the corner spinner, which now needs to be shared/ref-counted.

(Maybe we'll ultimately decide that you should have to click 'Update' to actually update those checkbox values? It is a little bit of an adjustment that the lead_item actions are immediate, while standard lead actions require clicking 'Update' to actually update.)