Add 'roll_call' parameter to filters for avoiding extra work #4337
Fixes the following:

Since Filter CO has to check all of its Filter CE facets to know if its base_query_ce_facet should be included in results, and then it needs to compute all of those facets again to actually get the filter result, this halves the amount of work done by each facet to which roll_call processing is added (when that facet is used in the filter query of course).

And since the previous sentence is repeated again in order to process data_attributes results in the `brief` action, this actually saves two rounds of processing per Filter CO call (again, for facets that use roll_call that are included in the filter result).

As an example, on my computer using the Stewart Island Gaz (from New Zealand ecoregions) to spatial filter COs, filter run time goes from 3m16s to 1m38s. That Gaz has 35,000 vertices and 99% of the work done in the spatial filter is done *before* the scope is returned from the spatial facet.