Don't show background spinner over Create Gaz modals
When we add a shape from, e.g., the draw-on-leaflet map modal, it updates the preview panel in the background - previously the spinner for that background update was obstructing the foreground modal.

It's possible this could lead to overlapping preview_shape requests, and that those overlapping requests could be returned out of order (I assume). I'm willing to live with that (for now).