Factor out encode to RGeo/decode to WKT/encode to RGeo loop from spatial filter processing #4337
For spatial filtering against a GA or GZ, we were taking a computed (union) shape from postgis, parsing/constructing an RGeo shape from that, immediately destructuring the RGeo shape back into text, passing the text to another function which immediately parsed the wkt back into an RGeo shape so that it could get the wkt's geometry type, and then passing the wkt and geometry type on.

*That's* what was making that spatial query so slow on large shapes, not the actual spatial filtering itself.

To continue from the last commit's example on Filter CO on the Fiordland gaz, the roll_call reduced that from 2 minutes to 1 minute, this fix reduces from 1 minute to 1.5 seconds. Yay!