Fix duplicate filter results on match_identifiers, take 1 #4323
This fixes the duplicate results but now, in the case where there *are* duplicate results, query.pluck(:id) is broken because the solution to the duplicate results was to add .select('DISTINCT ...'), whereas .pluck (from what I'm told) works by undoing the existing select and adding a new one.

Could do
.select("#{table.name}.*, sid.s")
.group("#{table.name}.id", 'sid.s')
.order('sid.s')
but that "breaks" .size (and .count) on the result set

Could we instead just add .distinct to the end of the query? Maybe?? The query has a select across a join because we need to preserve order of the results by match array order on identifiers. AFAICT .distinct in that case can be flaky in rails, in a way I don't understand well enough to use it.

On a similar note - which would be a better way to fix this bug if possible - see https://github.com/SpeciesFileGroup/taxonworks/blob/e1d952e23fd34f93fc620b0b7ecd2733196e2f12/lib/queries/query/filter.rb#L838-L839

Longer-winded and perhaps slower, but also perhaps more reliable:
deduped = target_query.joins(ids that match the input match_ids).distinct
ordered = target_query.joins(ids that match the input match_ids, along with the index of the id that matches).sort(index of id that matches).where(id: deduped)