LLM-fix issue with #id links causing page reloads in v-html'ed key #4287
More precisely:
print_key_markdown uses internal #id links to navigate between leads of the key. We red_carpet render the markdown to html and then v-html that in vue.
When you click on one of those href="#id" links it reloads the page. My vue-fu is not strong enough to know why.