Hook up taxon_name and taxon_name_relationship filters #4010
For now at least, in both directions, the connection is on both subject and object of relationships. Going *to* TaxonName there are facets to restrict the query down to just subject or just object, but in the other direction it's currently not possible.

(This took a long time to find all of the places I needed to fix the plumbing; maybe (more) documentation would be helpful.)