Fix async case of georef create for Filter CE #4401
STR:
1) Check a bunch (say > 25) of CEs in Filter CE.
2) Run the Set Georef task on them; it should run async.
3) Check your CEs for their new georefs, find no new ones.

The issue is that, although we're adding a `by:` value to the top level of the CE params we want to create on our filter results, we're *not* adding it to the individual georef_attributes array items, which is where it, along with project_id, is needed to create those georefs in the background.

```ruby
collecting_event: {
  georeferences_attributes: [
    {
      geographic_item_id: georef2.geographic_item.id,
      type: 'Georeference::VerbatimData',
    },
    {
      geographic_item_id: georef3.geographic_item.id,
      type: 'Georeference::Leaflet',
    }
  ],
  by: ...
}
```

To fix I updated QueryBatchRequest/query_batch_update to hold/use project_id and user_id as Current in async jobs. (We could pass the georef we're copying from's project_id, but not its update_by_id; and we could add both pieces of data in the controller to each embedded object; this solve both cases at once in a clean way I think(?)).