Finish getting Browse AD working for any object type #4025
There's still the open question of whether the query param for a polymorhpic type should be, e.g.,
```
&otu_id=1
```
or
```
&asserted_distribution_object_id=1&asserted_distribution_object_type='Otu'
```
Currently it's the former (I think this will match better with existing conventions on other tasks sending to this one), I kind of think it should actually be the latter (or either/both).

TODO: otu_id etc. should be added to the query param on each search, along with history state so that back/forward buttons work.