Clarify facet checkbox for or vs. all #3511
The UI label for the checkbox was 'Any', the param for it was (by convention) called *_id_or, but a value of true meant results had to match *all* of the facet items, which doesn't fit with "any" or "or".

I switched the UI label to 'All' and the params to *_id_all, where true means *all* facet terms must match.