Refactor base querying in Filter BA #4025
* *All* base object (i.e. otu, co) query params were being sent separately as subject query, object query, even though subject_taxon_name_id and object_taxon_name_id were the only ones that depended on subject/object. That meant that for all other params we were querying twice as much as needed.
* When the first point was in use, we were applying the taxon_name_mode param (which says to either `and` subject/object results or `or` them) to *all* of the base query search results, not just *taxon_name_id results. I think that was in error, though I haven't verified.

Now we have a separate taxon_name_id facet, because its result depends on and-ing or or-ing subject/object BA results.
Now the base query is only done once for Otu and for CO, and then joined to subject BAs and object BAs and returned as the union of those two.

If I put a byebug at the bottom of Filter.all, a spatial search prior to this pr would hit 18 times. Now it hits 11 times.