Always do && prefilter in covered_by_geographic_items_sql #4464
Unexepectedly to me, to /this/ function it makes very little timing difference either way to the large longitude-spanning shapes I looked at, but apparently (I guess) combined with subsequent queries the presence of the && prefilter can make the combination ~3 times slower.

E.g. /with/ the prefilter, a filter from the UI on United States is ~12 seconds, whereas without && it's ~4 seconds.
And Northern America from the UI is ~40 seconds /with/ the && prefilter, but ~15 seconds without it.