Don't fit to bounds in Leaflet while drawing a shape #4483
STR: In the New Gazetteer task, draw one shape. Now choose the draw line or polygon tool, zoom in near the existing shape, and start clicking inside the existing shape.
Eventually you should see the map zoom out so the entire original shape is visible.

I don't know why it doesn't happen on every click on the original shape, but this is because of the zoomToFeature set on each layer of the map, which runs on 'click' on every layer we add.

We now skip the zoomToFeature if any shape is being actively drawn.