Remove pagination before running Filter Taxon *ify queries #4494
As soon as `q` with pagination is used in a subquery in an *ify query (which they all do) you've restricted the ify query to only whatever page of results the original query returned.

In some cases it was probably never obvious because the ify query gets combined with the original result set, so result counts aren't so obviously low.