Handle download deleted during background job #2830
Awkward... Is it because the download is "ready" when a file exists at a certain predetermined-by-download-data location, which is a condition that cannot be atomically joined to the existence of db data without locking the db(?).

TODO: how many times is this job retrying when its download is already permanently gone? Is there any point to retrying in any case?