Change Obs. validation error string so Otus can be unified past matrices #4140
!! Close, but this doesn't quite work. Row items have no association with their rows, so when the second row item is unify-destroyed, it hasn't passed its row's associations to row 1. The citation still points to the now-nil row2. !!
Note order of relationship processing is alpha by lower snake case: 'observation_matrix_row_items', 'observation_matrix_rows', 'observations'. If rows were unified before row_items, this might work.

The case needed for is when observations on different otus being unified have the same character state value: we go to unify the otus, so we try to unify their related observations.
When two of those observations become the same after one of their observation_objects is changed, then we try to unify the observations, but *only* when the relation error matches the object to unify (and there are exactly 2 identical objects to unify).

Note that there are still simple conditions that will cause this unify to fail. If, for example, one of the rows is also referenced by a dynamic row item, then one row will have reference_count 1 and the other reference_count 2, and unifying those two objects will fail, causing the whole thing to fail.