Don't allow _delete for AD origin_citation #4514
This one was frustrating, I couldn't find a way to raise an error on AD.update! no matter what I tried in Citation's before_destroy, if the destroy was on the _origin_ citation and via nested _destroy.

LLM seemed to take both sides on this point, but it's perhaps because origin_citation is a has_one in the nested situation.