Update Cached Map status page, add AD cm scope
We've always only created cached_map_items for ADs that have shape in maps.rake, which has been one source of unregistered ADs. Now we also have ADs for otus with no taxon name (when processed, those are not registered), which complicates the numbers again. We may want to revisit those decisions just to simplify the counting/sanity checking.