Add Media section to DwCA meta.xml, add coreid column #1488
The TW-produced dwca with media extension now passes GBIF validation (for my small example).

I *think* the media section in meta.xml is required for extensions to be processed, though it's not 100% written that way (docs say meta.xml MAY contain 0 or more extension sections, doesn't say MUST contain 1 or more when there's actually an extension).

https://dwc.tdwg.org/text/

The linked docs *do* say that IF an extension element is included then:
`every extension entity MUST have an explicit <coreid> field to relate the extension record to a row in the core entity`