Connect Resource Relationship extension file to core (star-like) #1488
I believe this is necessary for GBIF etc. to actually display relationship data on an occurrence's page (otherwise extension records are just dropped I think).
Note we *want* to match multiple times to the core for a given relationship if both subject and object are in the core (at least one should be guaranteed to be there) - otherwise the relationship won't display on one or the other's page.

It probably doesn't seem like it, but the hardest part of doing this was recognizing that we needed to set project_id in the background job; otherwise queries fail in ways that take-me-longer-than-I-want-to-admit to fix (like filter#project_only? failing).