Further refine parent/child processing for 'has > 0 associations' #4514
The story:
All specs were passing as expected, but when I went to Browse Otu -> AD modal in object radial -> edit ad -> remove last citation, the controller was returning an exception.

I already had that exact spec, and its `update` was returning the expected `false`; controller's `update` was returning an exception :/

I tried to mimic the controller setup in spec (see below), but that never worked.

It was fixed by now skipping child processing when child is marked_for_destruction?, but we're still missing a spec that would raise if marked_for_destruction? were not in use.

+              # controller with this setup, where it raises RecordNotDestroyed -
+              # it's not working.
+              ad = AssertedDistribution.find(asserted_distribution.id)
+              expect(ad.citations.count).to eq(1)
+              params_hash = {
+                "asserted_distribution" => {
+                  "citations_attributes" => [
+                    { "id" => ad.citations.first.id, "_destroy" => true }
+                  ]
+                }
+              }
+              params = ActionController::Parameters.new(params_hash)
+              ad.update(params.require(:asserted_distribution).permit(citations_attributes: [:id, :is_original, :_destroy, :source_id, :pages, :citation_object_id, :citation_object_type]))
+              expect(ad.errors[:base].first).to match('citation is not provided')
+              expect(ad.reload.citations.count).to eq(1)