Require removal of all 'STUB' text from EML (one way or another) #2830
I'd rather not force people to remove it all before being able to partial save their EML, since likely many will have questions on how to fill in some of the fields (and so will want to save partial progress). But that means there are loopholes that allow STUB text on public downloads, so we 500 in that case.