Let computed filterDescriptors trump saved filterDescriptors #4527
STR without this patch (this is how it happened for me at least):
1. Send key otus to interactive key
2. Choose a (useful) state for an unused descriptor

Notice the number of Remaining descriptors doesn't change, and that when you look at the OM load in devtools, there's no param for descriptors.

This wasn't an issue before the previous patch existed; I haven't tracked down why.