[Tentative] Add AP creation to Origin Relationship quick form #4407
This would be setting a new precedent - no other OR types currently allow creation in this form.

The issue is: e.g., there is no association between CO and AP. Instead we have an origin relationship betweeen the two. (We probably don't want to maintain two lists?)
Quick forms can, by design, only be used to create new association objects. So technically we can't have an AP quick form slice since APs are (rails-)associated with no models.

So this is a compromise we may or may not want to proceed with. Currently the alternative would be to create the new AP from a New AP task or from the data New AP task, then search for that new AP in the AP search form on the quick form for Origin Relationship on a CO.