Refactor OriginRelationship quick form more, for draggable + delete #4407
This might still need work: only the 'new' table can be reordered, but keeping separate 'old' and 'new' lists with 'list = old + new' kind of breaks the expectations of `useSlice`.

Also the existing way of updating position is broken in some cases (probably in other components as well) due to the way acts_as_list works when you add an item that has the same position as an existing item.