Correct AP validations, add cached_otu_id #4407
I think it's 'fine', but this is clunky. AP requires an OriginRelationship chain that starts with AP and goes up to 'an otu', but that chain can't exist until the AP is created.

The solution here is to *only* create AP via a special call that verifies a chain without the AP, creates the AP, and then creates the originRelationship from the AP to the verified chain.