Permit destroy of terminal APs only #4407
Thanks to ai for pointing me to `OriginRelationshipDestroyContext < ActiveSupport::CurrentAttributes` (same way `Current` works) when I/we couldn't get `destroyed_by_association` to work in this context.

I changed destroy on AP to be disallowed unless it's the *terminal* object of the chain. It seemed like deleting an AP with child and Extract or Sequence would be a bad thing.

You can now delete APs from the graph using the navigator menu, and it will stay on the updated graph. (Otherwise you'd have to search for each object of the chain and delete one at a time.)