Reload leads after lead_item changes
Lead item changes can change lead otu; we want to keep ui lead data up-to-date, both for debugging and so that we can display the lead otu on the lead as it is (auto-magically) set.

That said, the lead otu should probably be read-only when there's an otu list on the lead.