Don't close context menu when clicking on autocomplete search results #4407
STR: Click on 'Create New AP' from the context menu of a node in Graph AP, search for an ontology term, click on a result in the results list, the context menu disappears.

I think autocomplete search results changed the way they were displayed recently, so the previous checks stopped working - another sign of what I already knew, which is that this is a bit flaky.

Maybe we could just check for click target ancestor with large z-value instead??