Refactor dwca/data#predicate_data to use a temp table #4607
With Claude.

In words, instead of a join where each co may appear in several rows, one for each da on that co, now each co appears in one row and there's a column for each da.
That allows us to directly read off the desired DA data for each co, and PG does all the work.