Add 'control panel' for complete downloads
Code by Claude.

Also try to be clearer about the meaning of max_age. This was designed when we were thinking about some projects getting polled by GBIF every day, or once a month, etc.

As it is, GBIF calls everybody once a week - if it's going to stay that way then we should probably get rid of that confusing setting and just set it internally.