api image link shortener optimization #4607
We shouldn't need to check each api image link one at a time to see if it's been shortened or not - here we check them all in one query.

This will mean each week when a new export is created, only the images that are new that week will need to get their own individual shortener call, all the rest will just be returned by the one query.

Downside: we're directly accessing shortener table columns by name. The columns we need (key is link you want to shorten, value is short-string for that link) seem very simple and basic, so personally I would expect those columns to never change, but that's no guarantee. (If they do change it most likely means we'll start blowing up in specs.)