Remove redundant indices added earlier in this branch #4607
The question was: is it better to use composite index (id, type) [exists already] or (type, id). For performance *using the composite index on both types* the answer in this case seems to be there's essentially no difference - here are results of a claude-generate benchmark:

```
Interesting! Now with (type, id) running first, the results flipped:

  Results (type, id first)

  (type, id) index: 10,671 queries/second
  - Average: 0.094ms per query
  - Total time: 9.371s

  (id, type) index: 11,610 queries/second ✓ FASTER
  - Average: 0.086ms per query
  - Total time: 8.613s

  The (id, type) index is ~9% faster

  This is the opposite of what we saw when (id, type) ran first! This suggests there might be cache warming effects or other system-level factors affecting the results.

  When we ran (id, type) first:
  - (id, type): 10,976 q/s
  - (type, id): 11,412 q/s (4% faster)

  When we ran (type, id) first:
  - (type, id): 10,671 q/s
  - (id, type): 11,610 q/s (9% faster)

  The second test always seems to perform better, regardless of which index it is! This indicates the benchmark might be measuring cache effects more than index efficiency.

  Conclusion

  The performance difference between the two index orders appears to be negligible - both are effectively the same speed when accounting for system variance and cache effects.
```

Where order *does* seem to matter is the extent to which your composite index can be used in less-composite cases: namely, the (id, type) index could be used as an index on id, the (type, id) index could be used as an index on type (and not vice versa): https://use-the-index-luke.com/sql/where-clause/the-equals-operator/concatenated-keys "The most important consideration when defining a concatenated index is how to choose the column order so it can be used as often as possible."