Don't try to .count on result with extra columns in Match CO -> update TD
STR: In the Match CO task or in the Match Identifiers facet of Filter CO, enter an ID that matches, then use the *left* object radial to add a taxon determination. Get the 500 below.

From a 500 report - when we do order_by: match_identifiers it adds an extra ordering position column to the result, which breaks .count in the batch query code. The extra position column is only used to order (which doesn't require it to be selected).

```
An ActiveRecord::StatementInvalid occurred in collection_objects#batch_update:

  PG::UndefinedFunction: ERROR:  function count(collection_objects, integer) does not exist
LINE 1: ...(sid.cached) IN ($1, $2, $3, $4, $5, $6))) SELECT COUNT(coll...
                                                             ^
HINT:  No function matches the given name and argument types. You might need to add explicit type casts.

  lib/query_batch_request.rb:88:in 'QueryBatchRequest#total_attempted'

-------------------------------
Github link:
-------------------------------

  https://github.com/SpeciesFileGroup/taxonworks/blob/ce6d689c7/lib/query_batch_request.rb#L88

-------------------------------
Request:
-------------------------------

  * URL        : https://sfg.taxonworks.org/collection_objects/batch_update.json
  * HTTP Method: PATCH
  * Parameters : {"collection_object_query" => {"match_identifiers_type" => "identifier", "match_identifiers" => "INHS Insect Collection 850610\nINHS Insect Collection 850606\nINHS Insect Collection 850607\nINHS Insect Collection 850608\nINHS Insect Collection 850609\nINHS Insect Collection 856213", "match_identifiers_delimiter" => "\n", "order_by" => "match_identifiers"}, "collection_object" => {"taxon_determinations_attributes" => [{"year_made" => "2025", "otu_id" => 169357, "roles_attributes" => [{"first_name" => "William G.", "last_name" => "Ruesink", "person_id" => 43029, "type" => "Determiner"}]}]}, "controller" => "collection_objects", "action" => "batch_update", "format" => "json"}
```