Fix several issues with parentage of infraspecific taxa #4643
It's not clear to me if these are always the right thing to do: the issues were cases where infraspecific taxa weren't getting parentID assigned, in the second case because the parent didn't exist in the known taxon list created up to that point. In that case we now construct the genus species parent of that taxon and add it to the export list. ??